more fixes but I doubt I'm gonna land this branch.
